+++ resources/views/templates/wrapper.blade.php
@@ -27,7 +27,7 @@
             @endif
             @if(!empty($siteConfiguration))
                 <script>
-                    window.SiteConfiguration = {!! json_encode($siteConfiguration) !!};
+                    window.SiteConfiguration = {!! json_encode($siteConfiguration, JSON_HEX_TAG | JSON_HEX_AMP | JSON_HEX_APOS | JSON_HEX_QUOT | JSON_UNESCAPED_UNICODE) !!};
                 </script>
             @endif
             @if(!empty($localeData))